refactor index extract logic to lib/handler
